crewai-backend/requirements.txt
fastapi
uvicorn
nats-py
python-dotenv
